Simulated data
Note: these are NOT the observed data used in the manuscript. These are random draws from a Poisson distribution, where the mean is the model-fitted median for the epi-year/serotype/agegroup/comorbidity level. Therefore, using these data with the provided code should give answers similar to, but not equal to, those found in the paper.